+++ src/search/Search.jsx
@@ -109,8 +109,13 @@ class Search extends React.Component {
          * 是否显示清除按钮
          */
         hasClear: PropTypes.bool,
+        /**
+         * 是否禁用
+         */
+        disabled: PropTypes.bool,
         locale: PropTypes.object,
         rtl: PropTypes.bool,
+
     };
 
     static defaultProps = {
@@ -125,6 +130,7 @@ class Search extends React.Component {
         onSearch: func.noop,
         onFilterChange: func.noop,
         hasClear: false,
+        disabled: false,
     };
 
     constructor(props) {
@@ -159,6 +165,9 @@ class Search extends React.Component {
     };
 
     onSearch = () => {
+        if (this.props.disabled) {
+            return;
+        }
         this.props.onSearch(this.state.value, this.state.filterValue);
     };
 
@@ -171,6 +180,9 @@ class Search extends React.Component {
     };
 
     onKeyDown = (e) => {
+        if (this.props.disabled) {
+            return;
+        }
         if (e.keyCode !== KEYCODE.ENTER) {
             return;
         }
@@ -178,7 +190,7 @@ class Search extends React.Component {
     }
     render() {
         const {
-            shape, filter, hasIcon,
+            shape, filter, hasIcon, disabled,
             placeholder, type, className,
             style, size, prefix, searchText,
             dataSource, filterProps, buttonProps,
@@ -200,13 +212,13 @@ class Search extends React.Component {
                 [`${prefix}search-icon`]: true,
                 [buttonProps.className]: !!buttonProps.className
             });
-            searchIcon = <Icon {...buttonProps} type="search" tabIndex="0" role="button" className={cls} onClick={this.onSearch} onKeyDown={this.onKeyDown}/>;
+            searchIcon = <Icon {...buttonProps} type="search" tabIndex="0" role="button" aria-disabled={disabled} className={cls} onClick={this.onSearch} onKeyDown={this.onKeyDown}/>;
         } else {
             const cls = classNames({
                 [`${prefix}search-btn`]: true,
                 [buttonProps.className]: !!buttonProps.className
             });
-            searchBtn = (<Button  {...buttonProps} tabIndex="0" className={cls} onClick={this.onSearch} onKeyDown={this.onKeyDown}>
+            searchBtn = (<Button  {...buttonProps} tabIndex="0" className={cls} onClick={this.onSearch} onKeyDown={this.onKeyDown} disabled={disabled}>
                 {hasIcon ? <Icon type="search" /> : null}
                 {searchText ? <span className={`${prefix}search-btn-text`}>{searchText}</span> : null}
             </Button>);
@@ -219,6 +231,7 @@ class Search extends React.Component {
                     hasBorder={false}
                     dataSource={filter}
                     size={size}
+                    disabled={disabled}
                     value={this.state.filterValue}
                     onChange={this.onFilterChange}
                 />
@@ -246,6 +259,7 @@ class Search extends React.Component {
                 value={this.state.value}
                 onChange={this.onChange}
                 popupContent={popupContent}
+                disabled={disabled}
             />
         </Group>);